refactor: ScriptState
